requirements.txt
streamlit
pandas
pickle
sklearn.preprocessing
numpy
sklearn.tree
sklearn.model_selection
sklearn.metrics